refactor(Scripts/ScarletEnclave): Consolidate chapter scripts into a single file. (#26804)
