--- src/server/scripts/EasternKingdoms/ScarletEnclave/zone_the_scarlet_enclave.cpp
@@ -1,143 +1,0 @@
-/*
- * This file is part of the AzerothCore Project. See AUTHORS file for Copyright information
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "CreatureScript.h"
-#include "PassiveAI.h"
-#include "Player.h"
-#include "ScriptedCreature.h"
-
-/*####
-## npc_valkyr_battle_maiden
-####*/
-
-enum Spells_VBM
-{
-    SPELL_REVIVE                = 51918
-};
-
-enum Says_VBM
-{
-    WHISPER_REVIVE              = 0
-};
-
-class npc_valkyr_battle_maiden : public CreatureScript
-{
-public:
-    npc_valkyr_battle_maiden() : CreatureScript("npc_valkyr_battle_maiden") { }
-
-    CreatureAI* GetAI(Creature* creature) const override
-    {
-        return new npc_valkyr_battle_maidenAI(creature);
-    }
-
-    struct npc_valkyr_battle_maidenAI : public PassiveAI
-    {
-        npc_valkyr_battle_maidenAI(Creature* creature) : PassiveAI(creature) { }
-
-        uint32 FlyBackTimer;
-        float x, y, z;
-        uint32 phase;
-
-        void Reset() override
-        {
-            me->setActive(true);
-            me->SetVisible(false);
-            me->SetUnitFlag(UNIT_FLAG_NON_ATTACKABLE);
-            me->SetCanFly(true);
-            FlyBackTimer = 500;
-            phase = 0;
-
-            me->GetPosition(x, y, z);
-            z += 4.0f;
-            x -= 3.5f;
-            y -= 5.0f;
-            me->GetMotionMaster()->Clear(false);
-            me->SetPosition(x, y, z, 0.0f);
-        }
-
-        void UpdateAI(uint32 diff) override
-        {
-            if (FlyBackTimer <= diff)
-            {
-                Player* player = nullptr;
-                if (me->IsSummon())
-                {
-                    if (Unit * summoner = me->ToTempSummon()->GetSummonerUnit())
-                    {
-                        player = summoner->ToPlayer();
-                    }
-                }
-
-                if (!player)
-                {
-                    phase = 3;
-                }
-
-                switch (phase)
-                {
-                    case 0:
-                        me->SetWalk(false);
-                        me->HandleEmoteCommand(EMOTE_STATE_FLYGRABCLOSED);
-                        FlyBackTimer = 500;
-                        break;
-                    case 1:
-                        if (player)
-                        {
-                            player->GetClosePoint(x, y, z, me->GetObjectSize());
-                        }
-                        z += 2.5f;
-                        x -= 2.0f;
-                        y -= 1.5f;
-                        me->GetMotionMaster()->MovePoint(0, x, y, z);
-                        if (player)
-                        {
-                            me->SetTarget(player->GetGUID());
-                        }
-                        me->SetVisible(true);
-                        FlyBackTimer = 4500;
-                        break;
-                    case 2:
-                        if (player && !player->isResurrectRequested())
-                        {
-                            me->HandleEmoteCommand(EMOTE_ONESHOT_CUSTOM_SPELL_01);
-                            DoCast(player, SPELL_REVIVE, true);
-                            Talk(WHISPER_REVIVE, player);
-                        }
-                        FlyBackTimer = 5000;
-                        break;
-                    case 3:
-                        me->SetVisible(false);
-                        FlyBackTimer = 3000;
-                        break;
-                    case 4:
-                        me->DisappearAndDie();
-                        break;
-                    default:
-                        //Nothing To DO
-                        break;
-                }
-                ++phase;
-            }
-            else FlyBackTimer -= diff;
-        }
-    };
-};
-
-void AddSC_the_scarlet_enclave()
-{
-    new npc_valkyr_battle_maiden();
-}
